Modified object rendering in infoview
